servicio-usuarios/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-decouple==3.8
google-cloud-secret-manager==2.18.2
